Added Item Interaction
Added an Interaction Interface that allows items to be interacted with by pressing the 'E' key, when within a certain range. If the player is in range, UI text will pop up with the key 'E'. Want to get item highlights to be apart of this interface, but for now it is separate